# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Seletivo Seplag Gov-MT",
        "version": "0.1"
    },
    "paths": {
...
        "/api/endereco": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Mostra os Endereços cadastrados",
                "description": "Lista os Endereços",
                "operationId": "9c10249b6d54a91b77e519f61d558174",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Nº da página",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Cadastra um novo Endereço",
                "description": "Registro de um novo Endereço",
                "operationId": "3768fdb817d263ccfb30b72455588fb5",
                "parameters": [
                    {
                        "name": "end_tipo_logradouro",
                        "in": "query",
                        "description": "Tipo logradouro",
                        "required": true
                    },
                    {
                        "name": "end_logradouro",
                        "in": "query",
                        "description": "Logradouro",
                        "required": true
                    },
                    {
                        "name": "end_numero",
                        "in": "query",
                        "description": "Número",
                        "required": true
                    },
                    {
                        "name": "end_bairro",
                        "in": "query",
                        "description": "Bairro",
                        "required": true
                    },
                    {
                        "name": "cid_id",
                        "in": "query",
                        "description": "Nº de identificação da Cidade",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/endereco/{end_id}": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Mostra um Endereco",
                "description": "Pesquisa Endereço através do (end_id)",
                "operationId": "6b6140ac69a6ca0ebccf35442a062420",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "Nº de identificação do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço Encontrado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Endereço não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Atualizar dados de um Endereço",
                "description": "Editar os dados de um Endereço através do (end_id)",
                "operationId": "8b02ed4414290f74a731935477be7b5b",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "Nº de identificação do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do Endereço atualizado com sucesso.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Erro ao atualizar o Endereço"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Exclui um Endereço",
                "description": "Exclui um Endereço através do (end_id)",
                "operationId": "aae5b72b297631baa40389a5846e6308",
                "parameters": [
                    {
                        "name": "end_id",
                        "in": "path",
                        "description": "ID do Endereço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Endereço excluído com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Não foi possível excluir o Endereço"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}